refactor: split server implementation into focused files
Move the package main implementation out of the single main.go file into concern-focused files for server lifecycle, roots, tool planning, execution, and watch handling while preserving behavior.

Refs #46

Refs #47
